reorg search-related classes to search namespace
