+++ spec/search/fielded_search_spec.rb
@@ -1,13 +1,13 @@
-describe FieldedSearch do
+describe Search::FieldedQuery do
   it "#to_s" do
-    fs = FieldedSearch.new({
+    fs = Search::FieldedQuery.new({
       foo: "bar"
     })
     expect(fs.to_s).to eq "foo:(bar)"
   end
 
   it "skips wildcard-only fields" do
-    fs = FieldedSearch.new({
+    fs = Search::FieldedQuery.new({
       foo: "bar",
       color: "*"
     })
@@ -15,19 +15,19 @@ describe FieldedSearch do
   end
 
   it "#present?" do
-    fs = FieldedSearch.new(nil)
+    fs = Search::FieldedQuery.new(nil)
     expect(fs.present?).to eq false
   end
 
   it "respects advanced value syntax" do
-    fs = FieldedSearch.new({
+    fs = Search::FieldedQuery.new({
       amount: ">100"
     })
     expect(fs.to_s).to eq "amount:>100"
   end
 
   it "#to_h" do
-    fs = FieldedSearch.new({
+    fs = Search::FieldedQuery.new({
       wild: "*",
       foo: "",
       bar: nil,
@@ -38,14 +38,14 @@ describe FieldedSearch do
   end
 
   it "#valid_for" do
-    fs = FieldedSearch.new({
+    fs = Search::FieldedQuery.new({
       amount: 123
     })
     expect(fs.value_for(:amount)).to eq 123
   end
 
   it "#humanized" do
-    fs = FieldedSearch.new({
+    fs = Search::FieldedQuery.new({
       "generator.name" => 123
     })
     expect(fs.humanized(Manifest).to_s).to eq "Generator name:(123)"
